Add tracking pixel
This is for an experiment to see how accurate google analytics numbers are. I
have it on the hack.wesleyac.com domain because I don't have access to the server
logs for GH pages - I'm not collecting any information other than what the default
nginx server log captures. Feel free to adblock it if it bothers you.

This is only going to be enabled for a short amount of time <1 month - if it's
been longer than that, open an issue and I'll remove it - I probably just forgot.